Fix: (highlight-function-calls--matcher)
This basically reverts 6c9c85c7998f7e199db99fe6997ac61bcc301d56 and
reimplements highlighting of function-quoted symbols more simply.  In
so doing it avoids the new bug which that change seemed to cause, in
which symbols immediately followed by a newline were not highlighted.

(In debugging that I seemed to discover bugs in Emacs's font-locking,
in which 1. the MATCHER function was not called when the match ended
at a newline, and 2. the MATCHER was being called at the same position
even when it returned non-nil; but since this works, I'm not pursuing
that further, because this has been hair-pulling enough.)